+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added BOM to allow easier usage
 - Added [JSpecify](https://jspecify.dev/) and [NullAway](https://github.com/uber/nullaway)
 - Added new [Crypto](crypto) module
+- Make event store capabilities explicit with [EventStoreCapabilities](api/src/main/java/org/fuin/esc/api/EventStoreCapabilities.java)
 
 ## 0.9.0
 - Added new `findAll()` method to [SerializedDataTypeRegistry](api/src/main/java/org/fuin/esc/api/SerializedDataTypeRegistry.java)